Started implementation of the Stripe embedded form
both backend and frontend -- successfully passed the PaymentDTO with all the details pertaining to a license plate to the OrderService

+++ backend/src/main/java/com/clockworkcode/sibiuairportparkingmanager/DTO/PaymentDTO.java
@@ -10,7 +10,7 @@ public class PaymentDTO {
     private Date endTime;
     private String airportCode;
     private String airportName;
-    private Float billableAmount;
+    private Long billableAmount;
 
     public PaymentDTO() {
     }
@@ -63,11 +63,11 @@ public class PaymentDTO {
         this.airportName = airportName;
     }
 
-    public Float getBillableAmount() {
+    public Long getBillableAmount() {
         return billableAmount;
     }
 
-    public void setBillableAmount(Float billableAmount) {
+    public void setBillableAmount(Long billableAmount) {
         this.billableAmount = billableAmount;
     }
 }
